refactor: get email on server side
